added down time page
